Get candidates non-promoter cis regulation

Get candidates non-promoter cis regulation

cis_em_genomic(
  meth_mat,
  expr_mat,
  gene_tss,
  min_samples = 100,
  max_k = 50,
  max_dist = 500000,
  min_dist = 200,
  spearman = FALSE,
  parallel = FALSE
)

Arguments

meth_mat
Matrix with methylation values. Each row is a locus and each column is a sample. Rownames should contain "chrom", "start" and "end" separated by "_"

expr_mat
Matrix with expression values. Each row is a gene and each column is a sample.

gene_tss
intervals set with additional "name" column with transcription start site (TSS) for each gene. Used for allocating a genomic coordinate for each gene. In case of duplicate names or coordinates the first one in the matrix will be used.

min_samples
minimal number of samples per gene (both expression and methylation). Default is 100.

max_k
maximal number of genes to report per locus. Default is 50. Note that for efficiency (memory and computation) it is best to keep this number small.

max_dist
maximal distance from TSS to consider "cis" (used for FDR computations)

min_dist
minimal distance from TSS (in order to exclude promoter regulation). Used for FDR computation.

spearman
use spearman correlation (if FALSE - use pearson)

parallel
use mclapply for computation of ranks. Note that for large matrices forking might fail due to lack of memory and therefore parallel should be set to FALSE.