Deconvolute TME (tumor microenvironment) effects from methylation data

Deconvolute TME (tumor microenvironment) effects from methylation data

deconv_TME(
  meth_mat,
  expr_mat,
  raw_meth_mat,
  min_meth = 0.1,
  max_meth = 0.9,
  min_expr = NULL,
  meth_cor_thresh = 0.25,
  expr_cor_thresh = 0.25,
  min_sd = NULL,
  k = NULL,
  k_meth = 30,
  k_expr = k_meth,
  caf_gene = "CAV1",
  immune_gene = "CD3D",
  scores_cor_thresh = 0.3,
  ...
)

Arguments

meth_mat
Matrix with methylation values to use for TME inference (usually - promoter methylation). Each row is a locus and each column is a sample. This can be a different matrix than raw_meth_mat.

expr_mat
Matrix with expression values. Each row is a gene and each column is a sample.

raw_meth_mat
Matrix with methylation values to deconvolute. Each row is a locus and each column is a sample. This can be a different matrix than meth_mat.

min_meth
minimal locus average methylation

max_meth
maximal locus average methylation

min_expr
minimal expression level. Only genes with at least one sample with expression above min_expr would be included.

meth_cor_thresh
minimal correlation level of a locus. Only loci with at least one correlation above the threshold will be included.

expr_cor_thresh
minimal correlation level of a gene. Only genes with at least one correlation above the threshold will be included.

min_sd
minimal standard deviation per locus.

k
k parameter for knn

k_meth
number of methylation clusters when clustering expression-methylation data

k_expr
number of expression clusters when clustering expression-methylation data

caf_gene
name of gene that is inside the CAF cluster (default: "CAV1")

immune_gene
name of gene that is inside the immune cluster (default: "CD3D")

scores_cor_thresh
minimal (anti) correlation of loci methylation to TME expression scores in order to create TME methylatoon scores

...
additional parameters to em_cross_cor

Value

a list with:

norm_methMatrix with normalized methylation values of raw_meth_mat

tme_featuresData frame with Immune and CAF expression and methylation signatures per sample

(tme_loci)list with intervals set of loci that were included in the immune and CAF methylation scores

em_crossList as returned from em_cross_cor. Mainly includes a matrix with expression-methylation correlation values that were used for the deconvolution algorithm, where rows are loci and columns are genes

em_cross_clustClustering of expression-methylation correlation data, as returned from cluster_em_cross_cor